runelite-client: update to substance 8.0.00-dev
